Add `ChainWorkerState::stage_block_execution`
Refactor to move the code from the `WorkerState` into the
`ChainWorkerState`.
